+++ src/Aws/Ecs/Exception/EcsException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Aws\Ecs\Exception;
+
+use Aws\Common\Exception\ServiceResponseException;
+
+/**
+ * Default service exception class
+ */
+class EcsException extends ServiceResponseException {}